refactor: remove redundant models
