+++ JWords/API/CoreDataClient.swift
@@ -454,131 +454,6 @@ class CoreDataClient {
 // MARK: Conversion
 extension CoreDataClient {
     
-    func checkIfExist(book: WordBook) throws -> Bool {
-        let fetchRequest = StudySetMO.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "title == %@", book.title)
-        
-        do {
-            return try !context.fetch(fetchRequest).isEmpty
-        } catch {
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-    }
-    
-    func convertBook(book: WordBook) throws {
-        guard let mo = NSEntityDescription.insertNewObject(forEntityName: "StudySet", into: context) as? StudySetMO else {
-            print("디버그: StudySetMO 객체를 만들 수 없음")
-            throw AppError.coreData
-        }
-        
-        mo.id = "set_" + UUID().uuidString + "_" + String(Int(Date().timeIntervalSince1970))
-        mo.title = book.title
-        mo.preferredFrontType = Int16(book.preferredFrontType.rawValue)
-        mo.isAutoSchedule = true
-        mo.closed = false
-        mo.createdAt = book.createdAt
-        
-        do {
-            try context.save()
-        } catch let error as NSError {
-            context.rollback()
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-    }
-    
-    func convert(input: ConversionInput, in set: StudySet) async throws -> StudyUnit {
-        
-        guard let mo = NSEntityDescription.insertNewObject(forEntityName: "StudyUnit", into: context) as? StudyUnitMO else {
-            print("디버그: StudyUnitMO 객체를 만들 수 없음")
-            throw AppError.coreData
-        }
-        
-        guard let set = try? context.existingObject(with: set.objectID) as? StudySetMO else {
-            print("디버그: objectID로 set 찾을 수 없음")
-            throw AppError.coreData
-        }
-        
-        if let kanjiImage = input.kanjiImage {
-            mo.kanjiImageID = try await iu.saveImage(data: kanjiImage)
-        }
-        
-        if let meaningImage = input.meaningImage {
-            mo.meaningImageID = try await iu.saveImage(data: meaningImage)
-        }
-        
-        HuriganaConverter.shared.extractKanjis(from: input.kanjiText)
-            .compactMap { try? getKanjiMO($0) }
-            .forEach { mo.addToKanjis($0) }
-        
-        mo.id = "unit_" + UUID().uuidString + "_" + String(Int(Date().timeIntervalSince1970))
-        mo.type = Int16(input.type.rawValue)
-        if !input.kanjiText.isEmpty { mo.kanjiText = input.kanjiText }
-        if !input.meaningText.isEmpty { mo.meaningText = input.meaningText }
-        mo.studyState = Int16(input.studyState.rawValue)
-        mo.createdAt = input.createdAt
-        mo.addToSet(set)
-        
-        do {
-            try context.save()
-            return StudyUnit(from: mo)
-        } catch let error as NSError {
-            context.rollback()
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-        
-    }
-    
-    func convert(inputs: [ConversionInput], in set: StudySet) async throws -> [StudyUnit] {
-        guard let set = try? context.existingObject(with: set.objectID) as? StudySetMO else {
-            print("디버그: objectID로 set 찾을 수 없음")
-            throw AppError.coreData
-        }
-        
-        var addedMO = [StudyUnitMO]()
-        
-        for input in inputs {
-            guard let mo = NSEntityDescription.insertNewObject(forEntityName: "StudyUnit", into: context) as? StudyUnitMO else {
-                print("디버그: StudyUnitMO 객체를 만들 수 없음")
-                throw AppError.coreData
-            }
-            
-            if let kanjiImage = input.kanjiImage {
-                mo.kanjiImageID = try await iu.saveImage(data: kanjiImage)
-            }
-            
-            if let meaningImage = input.meaningImage {
-                mo.meaningImageID = try await iu.saveImage(data: meaningImage)
-            }
-            
-            HuriganaConverter.shared.extractKanjis(from: input.kanjiText)
-                .compactMap { try? getKanjiMO($0) }
-                .forEach { mo.addToKanjis($0) }
-            
-            mo.id = "unit_" + UUID().uuidString + "_" + String(Int(Date().timeIntervalSince1970))
-            mo.type = Int16(input.type.rawValue)
-            if !input.kanjiText.isEmpty { mo.kanjiText = input.kanjiText }
-            if !input.meaningText.isEmpty { mo.meaningText = input.meaningText }
-            mo.studyState = Int16(input.studyState.rawValue)
-            mo.createdAt = input.createdAt
-            mo.addToSet(set)
-            
-            addedMO.append(mo)
-        }
-        
-        do {
-            try context.save()
-            return addedMO.map { StudyUnit(from: $0) }
-        } catch let error as NSError {
-            context.rollback()
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-        
-    }
-    
     func checkIfExist(_ kanjiText: String) throws -> StudyUnit? {
         
         if kanjiText.isEmpty { return nil }
@@ -624,30 +499,6 @@ extension CoreDataClient {
         
     }
     
-    func resetCoreData() throws {
-//        let setFetchRequest: NSFetchRequest<StudySetMO> = StudySetMO.fetchRequest()
-//        let unitFetchRequest: NSFetchRequest<StudyUnitMO> = StudyUnitMO.fetchRequest()
-//
-//        do {
-//            let sets = try context.fetch(setFetchRequest)
-//            for set in sets {
-//                context.delete(set)
-//            }
-//
-//            let units = try context.fetch(unitFetchRequest)
-//            for unit in units {
-//                context.delete(unit)
-//            }
-//
-//            try context.save()
-//
-//        } catch {
-//            context.rollback()
-//            NSLog("CoreData Error: %s", error.localizedDescription)
-//            throw AppError.coreData
-//        }
-//
-    }
 }
 
 // MARK: Interim function to convert kanjis from StudyUnit to StudyKanji
@@ -664,72 +515,6 @@ extension CoreDataClient {
         } catch {
             NSLog("CoreData Error: %s", error.localizedDescription)
             throw AppError.coreData
-        }
-    }
-    
-    private func convertKanjisToStudyKanji(kanji: StudyUnitMO) throws {
-        guard let kanjiText = kanji.kanjiText else {
-            print("디버그: kanjiText 없음")
-            return
-        }
-        
-        // 중복 검사
-        guard !(try! checkIfExist(kanji: kanjiText)) else {
-            print("디버그: 이미 DB에 존재하는 한자")
-            return
-        }
-        
-        // StudyUnitMO 찾기
-        guard let unitMO = try? context.existingObject(with: kanji.objectID) as? StudyUnitMO else {
-            print("디버그: objectID로 unit 찾을 수 없음")
-            throw AppError.coreData
-        }
-        
-        // mo에 넣기
-        guard let studyKanjimo = NSEntityDescription.insertNewObject(forEntityName: "StudyKanji", into: context) as? StudyKanjiMO else {
-            throw AppError.coreData
-        }
-        
-        let wikiKanji = KanjiWikiClient.shared.getWikiKanji(kanjiText)
-        
-        studyKanjimo.id = "kanji_" + UUID().uuidString + "_" + String(Int(Date().timeIntervalSince1970))
-        studyKanjimo.kanji = kanji.kanjiText
-        studyKanjimo.meaning = kanji.meaningText
-        studyKanjimo.ondoku = wikiKanji?.ondoku ?? ""
-        studyKanjimo.kundoku = wikiKanji?.kundoku ?? ""
-        studyKanjimo.createdAt = kanji.createdAt
-        
-        // mo에 unit과 연결
-        let words = unitMO.sampleForKanji ?? []
-        studyKanjimo.addToWords(words)
-        
-        do {
-            try context.save()
-            print("디버그) 이동한 kanji: \(studyKanjimo.kanji!)")
-        } catch let error as NSError {
-            context.rollback()
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-    }
-    
-    private func fetchOldKanjis() throws -> [StudyUnitMO] {
-        let fetchRequest = StudyUnitMO.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "type == \(UnitType.kanji.rawValue)")
-        
-        do {
-            return try context.fetch(fetchRequest)
-        } catch {
-            NSLog("CoreData Error: %s", error.localizedDescription)
-            throw AppError.coreData
-        }
-    }
-    
-    func convertAllKanjisToStudyKanji() throws {
-        let oldKanjis = try! fetchOldKanjis()
-        
-        for kanji in oldKanjis {
-            try! convertKanjisToStudyKanji(kanji: kanji)
         }
     }
