+++ JWords/Model/InputImageType.swift
@@ -1,0 +1,14 @@
+//
+//  InputImageType.swift
+//  JWords
+//
+//  Created by JW Moon on 2023/10/02.
+//
+
+#if os(iOS)
+import UIKit
+typealias InputImageType = UIImage
+#elseif os(macOS)
+import Cocoa
+typealias InputImageType = NSImage
+#endif
