+++ JWords/Model/StudyState.swift
@@ -1,0 +1,10 @@
+//
+//  StudyState.swift
+//  JWords
+//
+//  Created by JW Moon on 2023/10/02.
+//
+
+enum StudyState: Int {
+    case undefined = 0, success, fail
+}
