+++ JWords/Model/WordBookSchedule.swift
@@ -1,0 +1,10 @@
+//
+//  WordBookSchedule.swift
+//  JWords
+//
+//  Created by JW Moon on 2023/10/02.
+//
+
+enum WordBookSchedule: Equatable, CaseIterable {
+    case none, study, review
+}
