+++ JWords/iOSView/WordList/Cells/StudyCell.swift
@@ -72,8 +72,6 @@ struct StudyWord: ReducerProtocol {
     }
     
     private let cd = CoreDataClient.shared
-    @Dependency(\.cdWordClient) var wordClient
-    private enum UpdateStudyStateID {}
     
     var body: some ReducerProtocol<State, Action> {
         Reduce { state, action in
@@ -84,13 +82,13 @@ struct StudyWord: ReducerProtocol {
             case .cellDoubleTapped:
                 if state.isLocked { return .none }
                 state.studyState = .undefined
-                _ = try! wordClient.studyState(state.unit, .undefined)
+                _ = try! cd.updateStudyState(unit: state.unit, newState: .undefined)
                 return .none
             case .cellDrag(let direction):
                 if state.isLocked { return .none }
                 let newState: StudyState = direction == .left ? .success : .fail
                 state.studyState = newState
-                _ = try! wordClient.studyState(state.unit, newState)
+                _ = try! cd.updateStudyState(unit: state.unit, newState: newState)
                 return .none
             case .kanjiButtonTapped:
                 if state.kanjis.isEmpty {
